style: Apply clang-tidy standardizations
Automatically format syntax and enforce base-level readability rules.

- Standardize local variables and private members to Google style
- Update all referencing call sites and headers across the workspace
- Standardize local variables and private members to Google style
- Update all referencing call sites and headers across the workspace

+++ emulator/videobridge/media/in_process_audio_source.cc
@@ -23,11 +23,11 @@ InProcessAudioSource::InProcessAudioSource(uint32_t sample_rate, uint32_t channe
         : sample_rate_(sample_rate)
         , channels_(channels)
         , samples_per_10ms_channel_(sample_rate / 100)
-        , samples_per_10ms_total_((sample_rate / 100) * channels)
+        , samples_per_10ms_total_(static_cast<size_t>(sample_rate / 100) * channels)
         , temp_frame_buffer_(samples_per_10ms_total_) {
-    DCHECK_GT(sample_rate, 0u) << "Audio sample rate must be greater than 0.";
-    DCHECK_EQ(sample_rate % 100, 0u) << "Audio sample rate must be a multiple of 100 Hz.";
-    DCHECK_GT(channels, 0u) << "Audio channels must be greater than 0.";
+    DCHECK_GT(sample_rate, 0U) << "Audio sample rate must be greater than 0.";
+    DCHECK_EQ(sample_rate % 100, 0U) << "Audio sample rate must be a multiple of 100 Hz.";
+    DCHECK_GT(channels, 0U) << "Audio channels must be greater than 0.";
 }
 
 InProcessAudioSource::~InProcessAudioSource() {
